Moved InMemoryUserRepository to repository folder.
