+++ src/main/java/recipe/foodbar/app/manual/User.java
@@ -3,7 +3,7 @@ package recipe.foodbar.app.manual;
 import recipe.foodbar.controller.user.AccountController;
 import recipe.foodbar.controller.user.AccountPresenter;
 import recipe.foodbar.id_generator.jug.JugIdGenerator;
-import recipe.foodbar.repository.InMemoryUserRepository;
+import recipe.foodbar.repository.simpleDB.InMemoryUserRepository;
 import recipe.foodbar.usecase.user.UserInputData;
 import recipe.foodbar.usecase.user.UserManager;
 import recipe.foodbar.usecase.user.port.UserCreatorInputBoundary;
