+++ src/main/java/recipe/foodbar/config/JavelinConfig.java
@@ -2,7 +2,7 @@ package recipe.foodbar.config;
 
 import recipe.foodbar.encoder.sha256.Sha256PasswordEncoder;
 import recipe.foodbar.id_generator.jug.JugIdGenerator;
-import recipe.foodbar.repository.simpleDB.InMemoryUserRepository;
+import recipe.foodbar.repository.simpleDB.InMemoryUserExampleRepository;
 import recipe.foodbar.usecase.user_example.CreateUser;
 import recipe.foodbar.usecase.user_example.FindUser;
 import recipe.foodbar.usecase.user_example.LoginUser;
@@ -11,7 +11,7 @@ import recipe.foodbar.usecase.user_example.port.PasswordEncoder;
 import recipe.foodbar.usecase.user_example.port.UserRepository;
 
 public class JavelinConfig {
-    private final UserRepository userRepository = new InMemoryUserRepository();
+    private final UserRepository userRepository = new InMemoryUserExampleRepository();
     private final IdGenerator idGenerator = new JugIdGenerator();
     private final PasswordEncoder passwordEncoder = new Sha256PasswordEncoder();
     private final CreateUser createUser = new CreateUser(userRepository, passwordEncoder, idGenerator);
